Do not create a 'this' binding for global functions at all (except for indirect eval)
Only bind 'super' as a call target for super() inside class
constructors.

Skip throwing a SyntaxError when defer-parsing 'new.target' and we find
the enclosing non-lambda is global. If the 'new.target' is located in a
deferred-lambda, the enclosing non-lambda will always be null. We can
safely ignore the check here in deferred-parse mode because we would
have already thrown a SyntaxError in the upfront pass and we would not
be defer-parsing a lambda which had a SyntaxError in it to begin with.