Fix defer parsed lambda with default arguments containing function expressions
When we defer parsed a lambda, we would still go back and parse the lambda formal parameters twice assuming at first that they might not be lambda parameters. This is stilly, though, as we already parsed the lambda and we don't defer parse comma statements. So instead, we now just start parsing the function directly via `ParseFncDecl` whenever we're defer parsing a function. This way we can skip parsing the entire statement list (of which there is only one - the function we already verified syntax for) even for ordinary functions.