Fix lambda capturing arguments symbol
Lambdas don't have their own arguments binding but they capture the outer function. We previously did not save the arguments symbol information when we saved a `ScopeInfo` because functions could never reference the arguments symbol of their parent but lambdas can so we need to save and restore the symbol information for the arguments symbol.